# Conflicts:
#	Source/Applications/Backend/WCF/VodovozDeliveryRulesService/Program.cs
#	Source/Applications/Backend/WCF/VodovozDeliveryRulesService/VodovozDeliveryRulesService.csproj
#	Source/Applications/Desktop/Vodovoz/CreateProjectParam.cs
#	Source/Applications/Desktop/Vodovoz/Vodovoz.csproj
#	Source/Libraries/External/QSProjects
#	Source/Vodovoz.sln

+++ Source/Libraries/Core/Business/VodovozBusiness/EntityRepositories/Store/IWarehouseRepository.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using QS.DomainModel.UoW;
 using Vodovoz.Domain.Goods;
+using Vodovoz.Domain.Permissions.Warehouses;
 using Vodovoz.Domain.Store;
 
 namespace Vodovoz.EntityRepositories.Store
@@ -13,7 +14,7 @@ namespace Vodovoz.EntityRepositories.Store
 		IEnumerable<NomanclatureStockNode> GetWarehouseNomenclatureStock(IUnitOfWork uow, int warehouseId, IEnumerable<int> nomenclatureIds);
 		IEnumerable<Nomenclature> GetDiscrepancyNomenclatures(IUnitOfWork uow, int warehouseId);
 		bool WarehouseByMovementDocumentsNotificationsSubdivisionExists(IUnitOfWork uow, int subdivisionId);
-		decimal GetTotalShippedKgByWarehousesAndProductGroups(
+		int GetTotalShippedKgByWarehousesAndProductGroups(
 			IUnitOfWork uow, DateTime dateFrom, DateTime dateTo, IEnumerable<int> productGroupsIds, IEnumerable<int> warehousesIds);
 	}
